Update: Restrict FloatingChat to About, Themes, and ThemeDetail pages
